updated UserController

+++ src/main/java/com/mnilga/travel/agency/application/controller/UserController.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -23,32 +24,53 @@ public class UserController {
     }
 
     @PostMapping
-    public ResponseEntity<UserDto> create(@RequestBody User user){
+    public ResponseEntity<UserDto> create(@RequestBody @Valid User user){
+        if(user == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         UserDto createdUserDto = userService.create(user);
         return new ResponseEntity<>(createdUserDto, HttpStatus.CREATED);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<UserDto> readById(@PathVariable UUID id){
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         UserDto userDto = userService.readById(id);
+        if(userDto == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         return new ResponseEntity<>(userDto, HttpStatus.OK);
     }
 
     @PutMapping
-    public ResponseEntity<UserDto> update(@RequestBody User user){
+    public ResponseEntity<UserDto> update(@RequestBody @Valid User user){
+        if(user == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         UserDto updatedUser = userService.update(user);
         return new ResponseEntity<>(updatedUser, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<UserDto> delete(@PathVariable UUID id){
+        UserDto userDto = userService.readById(id);
+        if(userDto == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         userService.delete(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @GetMapping
     public ResponseEntity<List<UserDto>> getAllUsers(){
         List<UserDto> userDtoList = userService.getAllUsers();
+        if(userDtoList.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<>(userDtoList, HttpStatus.OK);
     }
